Implement cloud-config runcmd
If the userdata is of type cloud-config, the runcmd entry can contain
multiple entries with commands that will be executed, in the order
of their definition.

The commands can be given as a string or as an array of strings, the
first item being the binary to be executed and the rest being the
parameters of that binary.

The commands will be aggregated and written into one single shell file,
in the order of their definition.
On Windows, the file will be executed by the native Windows
shell cmd.exe.

Example userdata file:

runcmd:
 - 'dir C:\\'
 - ['echo', '1']

Fixes: https://github.com/cloudbase/cloudbase-init/issues/27

Change-Id: Ie307e08f8c4108c7bf9108543cc90b6a7fa2e7ae